+++ src/main/java/com/github/theredbrain/combatrollextension/CombatRollExtension.java
@@ -1,23 +1,15 @@
 package com.github.theredbrain.combatrollextension;
 
 import com.github.theredbrain.combatrollextension.config.ServerConfig;
-import com.github.theredbrain.combatrollextension.config.ServerConfigWrapper;
 import com.github.theredbrain.combatrollextension.entity.DuckLivingEntityMixin;
 import com.github.theredbrain.staminaattributes.entity.StaminaUsingEntity;
-import com.google.gson.Gson;
-import me.shedaniel.autoconfig.AutoConfig;
-import me.shedaniel.autoconfig.serializer.JanksonConfigSerializer;
-import me.shedaniel.autoconfig.serializer.PartitioningSerializer;
+import me.fzzyhmstrs.fzzy_config.api.ConfigApiJava;
 import net.combat_roll.api.RollInvulnerable;
 import net.combat_roll.api.event.ServerSideRollEvents;
 import net.fabricmc.api.ModInitializer;
-import net.fabricmc.fabric.api.networking.v1.PayloadTypeRegistry;
-import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents;
-import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
+import net.fabricmc.loader.api.FabricLoader;
+import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.attribute.EntityAttribute;
-import net.minecraft.network.RegistryByteBuf;
-import net.minecraft.network.codec.PacketCodec;
-import net.minecraft.network.packet.CustomPayload;
 import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.util.Identifier;
 import org.slf4j.Logger;
@@ -26,29 +18,36 @@ import org.slf4j.LoggerFactory;
 public class CombatRollExtension implements ModInitializer {
 	public static final String MOD_ID = "combatrollextension";
 	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
-	public static ServerConfig serverConfig;
+	public static ServerConfig SERVER_CONFIG;
 
 	public static RegistryEntry<EntityAttribute> ROLL_STAMINA_COST;
 	public static RegistryEntry<EntityAttribute> ROLL_INVULNERABLE_TICKS;
 
+	public static final boolean isStaminaAttributesLoaded = FabricLoader.getInstance().isModLoaded("staminaattributes");
+
+	public static float getCurrentStamina(LivingEntity livingEntity) {
+		float currentStamina = 0.0F;
+		if (isStaminaAttributesLoaded) {
+			currentStamina = ((StaminaUsingEntity) livingEntity).staminaattributes$getStamina();
+		}
+		return currentStamina;
+	}
+
+	public static void addStamina(LivingEntity livingEntity, float amount) {
+		if (isStaminaAttributesLoaded) {
+			((StaminaUsingEntity) livingEntity).staminaattributes$addStamina(amount);
+		}
+	}
+
 	@Override
 	public void onInitialize() {
 		LOGGER.info("Combat Roll was extended!");
-
-		// Config
-		AutoConfig.register(ServerConfigWrapper.class, PartitioningSerializer.wrap(JanksonConfigSerializer::new));
-		serverConfig = ((ServerConfigWrapper) AutoConfig.getConfigHolder(ServerConfigWrapper.class).getConfig()).server;
-
-		// Events
-		PayloadTypeRegistry.playS2C().register(ServerConfigSyncPacket.PACKET_ID, ServerConfigSyncPacket.PACKET_CODEC);
-		ServerPlayConnectionEvents.JOIN.register((handler, sender, server) -> {
-			ServerPlayNetworking.send(handler.player, new ServerConfigSyncPacket(serverConfig));
-		});
+		SERVER_CONFIG = ConfigApiJava.registerAndLoadConfig(ServerConfig::new);
 
 		ServerSideRollEvents.PLAYER_START_ROLLING.register((serverPlayerEntity, vec3d) -> {
 			if (!serverPlayerEntity.isCreative()) {
-				((StaminaUsingEntity) serverPlayerEntity).staminaattributes$addStamina(-(((DuckLivingEntityMixin) serverPlayerEntity).combatrollextension$getRollStaminaCost() * serverConfig.global_rolling_stamina_cost_multiplier));
-				if (serverConfig.is_roll_invulnerable_ticks_attribute_active) {
+				CombatRollExtension.addStamina(serverPlayerEntity, -(((DuckLivingEntityMixin) serverPlayerEntity).combatrollextension$getActualRollStaminaCost()));
+				if (SERVER_CONFIG.is_roll_invulnerable_ticks_attribute_active) {
 					((RollInvulnerable) serverPlayerEntity).setRollInvulnerableTicks((int) ((DuckLivingEntityMixin) serverPlayerEntity).combatrollextension$getRollInvulnerableTicks());
 				}
 			}
@@ -56,26 +55,7 @@ public class CombatRollExtension implements ModInitializer {
 
 	}
 
-	public record ServerConfigSyncPacket(ServerConfig serverConfig) implements CustomPayload {
-		public static final CustomPayload.Id<ServerConfigSyncPacket> PACKET_ID = new CustomPayload.Id<>(identifier("server_config_sync"));
-		public static final PacketCodec<RegistryByteBuf, ServerConfigSyncPacket> PACKET_CODEC = PacketCodec.of(ServerConfigSyncPacket::write, ServerConfigSyncPacket::new);
-
-		public ServerConfigSyncPacket(RegistryByteBuf registryByteBuf) {
-			this(new Gson().fromJson(registryByteBuf.readString(), ServerConfig.class));
-		}
-
-		private void write(RegistryByteBuf registryByteBuf) {
-			registryByteBuf.writeString(new Gson().toJson(serverConfig));
-		}
-
-		@Override
-		public CustomPayload.Id<? extends CustomPayload> getId() {
-			return PACKET_ID;
-		}
-	}
-
 	public static Identifier identifier(String path) {
 		return Identifier.of(MOD_ID, path);
 	}
-
 }