+++ src/main/java/com/github/theredbrain/combatrollextension/entity/DuckLivingEntityMixin.java
@@ -3,4 +3,5 @@ package com.github.theredbrain.combatrollextension.entity;
 public interface DuckLivingEntityMixin {
 	float combatrollextension$getRollInvulnerableTicks();
 	float combatrollextension$getRollStaminaCost();
+	float combatrollextension$getActualRollStaminaCost();
 }